Convert Reviews Section in home page to ssr
